requirements.txt
numpy
h5py
matplotlib
scikit-image
Pillow
cython
psutil
tqdm